SCP-3018: Merged modules
