+++ marlowe-actus/marlowe-actus.cabal
@@ -64,7 +64,6 @@ library
     Language.Marlowe.ACTUS.Generator.MarloweCompat
     Language.Marlowe.ACTUS.Model.POF.Payoff
     Language.Marlowe.ACTUS.Model.STF.StateTransition
-    Language.Marlowe.ACTUS.Model.SCHED.ContractScheduleModel
     Language.Marlowe.ACTUS.Model.SCHED.ContractSchedule
     Language.Marlowe.ACTUS.Model.INIT.StateInitializationModel
     Language.Marlowe.ACTUS.Model.APPL.Applicability
